refactor(tf): Use singleton class
Fixes #214
