+++ mlf_core/create/templates/common_mlflow_files/{{ cookiecutter.common_mlflow }}/{{ cookiecutter.project_slug_no_hyphen }}/mlf_core/mlf_core.py
@@ -7,6 +7,9 @@ from rich import print
 {%- if cookiecutter.language == "pytorch" -%}
 import torch
 import mlflow.pytorch
+{%- if cookiecutter.language == "tensorflow" -%}
+# ONLY FOR TENSORFLOW -> else RuntimeError: make_default_context() wasn't able to create a context on any of the 1 detected devices
+from system_intelligence.query import query_and_export  # noqa F401
 {% endif %}
 
 
@@ -53,6 +56,12 @@ class MLFCore:
         if num_gpus > 0:
             torch.cuda.manual_seed(seed)
             torch.cuda.manual_seed_all(seed)  # For multiGPU
+{%- elif cookiecutter.language == "tensorflow" -%}
+    def set_tensorflow_random_seeds(seed):
+        tf.random.set_seed(seed)
+        tf.config.threading.set_intra_op_parallelism_threads = 1  # CPU only
+        tf.config.threading.set_inter_op_parallelism_threads = 1  # CPU only
+        os.environ['TF_DETERMINISTIC_OPS'] = '1'
 {% endif %}
 
     def log_sys_intel_conda_env(self):
