+++ mlf_core/create/templates/mlflow/mlflow_tensorflow/{{ cookiecutter.project_slug }}/{{ cookiecutter.project_slug_no_hyphen }}/{{ cookiecutter.project_slug_no_hyphen }}.py
@@ -6,7 +6,7 @@ import os
 import time
 from rich import traceback, print
 
-from mlf_core.mlf_core import log_sys_intel_conda_env, set_general_random_seeds
+from mlf_core.mlf_core import MLFCore
 from data_loading.data_loader import load_train_test_data
 from model.model import create_model
 from training.train import train, test
@@ -67,8 +67,8 @@ def start_training():
         mlflow.tensorflow.autolog()
 
         # Fix all random seeds and Tensorflow specific reproducibility settings
-        set_general_random_seeds(dict_args["general_seed"])
-        set_tensorflow_random_seeds(dict_args["tensorflow_seed"])
+        MLFCore.set_general_random_seeds(dict_args["general_seed"])
+        MLFCore.set_tensorflow_random_seeds(dict_args["tensorflow_seed"])
 
         # Use Mirrored Strategy for multi GPU support
         strategy = tf.distribute.MirroredStrategy()
@@ -94,16 +94,9 @@ def start_training():
             print(f'[bold green]{device} Run Time: {str(time.time() - runtime)} seconds')
 
             # Log hardware and software
-            log_sys_intel_conda_env()
+            MLFCore.log_sys_intel_conda_env()
 
             print(f'[bold blue]\nLaunch TensorBoard with:\ntensorboard --logdir={os.path.join(mlflow.get_artifact_uri(), "tensorboard_logs", "train")}')
-
-
-def set_tensorflow_random_seeds(seed):
-    tf.random.set_seed(seed)
-    tf.config.threading.set_intra_op_parallelism_threads = 1  # CPU only
-    tf.config.threading.set_inter_op_parallelism_threads = 1  # CPU only
-    os.environ['TF_DETERMINISTIC_OPS'] = '1'
 
 
 if __name__ == '__main__':
